move hooks outside common folder
