+++ web/ui/react-app/src/pages/alerts/Alerts.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { RouteComponentProps } from '@reach/router';
 import PathPrefixProps from '../../PathPrefixProps';
-import { useFetch } from '../../common/hooks/useFetch';
+import { useFetch } from '../../hooks/useFetch';
 import { withStatusIndicator } from '../../common/withStatusIndicator';
 import AlertsContent, { RuleStatus, AlertsProps } from './AlertContents';
