+++ web/ui/react-app/src/pages/targets/ScrapePoolPanel.tsx
@@ -7,7 +7,7 @@ import EndpointLink from './EndpointLink';
 import TargetLabels from './TargetLabels';
 import { formatRelative, humanizeDuration } from '../../utils/timeFormat';
 import { now } from 'moment';
-import { useLocalStorage } from '../../common/hooks/useLocalStorage';
+import { useLocalStorage } from '../../hooks/useLocalStorage';
 import { ToggleMoreLess } from '../../common/ToggleMoreLess';
 
 interface PanelProps {
