+++ web/ui/react-app/src/pages/tsdbStatus/TSDBStatus.tsx
@@ -3,7 +3,7 @@ import { RouteComponentProps } from '@reach/router';
 import { Alert, Table } from 'reactstrap';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSpinner } from '@fortawesome/free-solid-svg-icons';
-import { useFetch } from '../../common/hooks/useFetch';
+import { useFetch } from '../../hooks/useFetch';
 import PathPrefixProps from '../../PathPrefixProps';
 
 export interface Stats {
